fix: resolve clippy warnings (#1165) (#1196)
* fix: resolve clippy warnings (#1165)

- Remove unused imports (HashSet, SinkExt, StreamExt, AgentEvent, etc.)
- Fix unused variables with underscore prefix
- Add #[allow(dead_code)] for test fixtures and future-use fields
- Replace split().last() with rsplit().next() for efficiency
- Replace map_or with is_some_and for clarity
- Remove unnecessary u32 cast
- Replace or_insert_with(Vec::new) with or_default()
- Remove identical if branches
- Use strip_prefix instead of manual prefix stripping
- Replace manual range check with RangeInclusive::contains
- Replace len() >= 1 with !is_empty()
- Use io::Error::other() instead of io::Error::new(ErrorKind::Other, _)
- Extract complex type into PermissionCheckFn alias
- Fix doc list item indentation
- Fix cfg condition: integration_tests -> integration-tests



* style: apply cargo fmt



---------

+++ crates/miyabi-web-api/src/routes/health.rs
@@ -43,7 +43,7 @@ pub async fn health_check(State(state): State<AppState>) -> (StatusCode, Json<He
     let database = match sqlx::query("SELECT 1").fetch_one(&state.db).await {
         Ok(_) => {
             let pool_options = state.db.options();
-            let total_connections = state.db.size() as u32;
+            let total_connections = state.db.size();
             let idle = state.db.num_idle() as u32;
             Some(DatabaseStats {
                 active_connections: total_connections.saturating_sub(idle),
